refact/front/BtnNextIcon.jsx: moves to src/components/buttons/BtnNextIcon.jsx
