+++ frontend/src/pages/trips/new/TripOriginForm.jsx
@@ -4,7 +4,7 @@ import { setNavigationLayout } from '../../../store/slices/navigationSlice';
 import { setOrigin } from '../../../store/slices/tripsSlice';
 import Input from "../../../components/Input"
 import volante from "../../../assets/volante.png"
-import BtnNextIcon from "../../../components/BtnNextIcon"
+import BtnNextIcon from "../../../components/buttons/BtnNextIcon"
 
 const TripOriginForm = () => {
   const dispatch = useDispatch();
